refactor: move zone_lifecycle + zone_finalizers from services/ to system_services/lifecycle/
Zone lifecycle is booted in _boot_system_services and belongs in the
SYSTEM tier alongside brick_lifecycle and brick_reconciler. Move the
module and its finalizer subpackage to their canonical location, update
all imports in factory/_system.py, system_services/lifecycle/__init__.py,
and test files.

- git mv 7 files from services/ to system_services/lifecycle/
- Update 5 internal imports in zone_finalizers/__init__.py
- Update 2 lazy imports in factory/_system.py
- Add ZoneLifecycleService re-export to lifecycle/__init__.py
- Update imports + patch target in 3 test files
- ruff/mypy clean, all 59 unit tests pass
