+++ src/nexus/system_services/lifecycle/zone_finalizers/__init__.py
@@ -1,0 +1,21 @@
+"""Zone finalizers — concrete cleanup services for zone deprovisioning (Issue #2061).
+
+Each finalizer implements ``ZoneFinalizerProtocol`` and handles one
+domain of zone-scoped resources (cache, search, mounts, ReBAC, bricks).
+"""
+
+from nexus.system_services.lifecycle.zone_finalizers.brick_drain_finalizer import (
+    BrickDrainFinalizer,
+)
+from nexus.system_services.lifecycle.zone_finalizers.cache_finalizer import CacheZoneFinalizer
+from nexus.system_services.lifecycle.zone_finalizers.mount_finalizer import MountZoneFinalizer
+from nexus.system_services.lifecycle.zone_finalizers.rebac_finalizer import ReBACZoneFinalizer
+from nexus.system_services.lifecycle.zone_finalizers.search_finalizer import SearchZoneFinalizer
+
+__all__ = [
+    "BrickDrainFinalizer",
+    "CacheZoneFinalizer",
+    "MountZoneFinalizer",
+    "ReBACZoneFinalizer",
+    "SearchZoneFinalizer",
+]
