+++ tests/unit/services/test_zone_finalizers.py
@@ -12,11 +12,13 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from nexus.services.zone_finalizers.brick_drain_finalizer import BrickDrainFinalizer
-from nexus.services.zone_finalizers.cache_finalizer import CacheZoneFinalizer
-from nexus.services.zone_finalizers.mount_finalizer import MountZoneFinalizer
-from nexus.services.zone_finalizers.rebac_finalizer import ReBACZoneFinalizer
-from nexus.services.zone_finalizers.search_finalizer import SearchZoneFinalizer
+from nexus.system_services.lifecycle.zone_finalizers.brick_drain_finalizer import (
+    BrickDrainFinalizer,
+)
+from nexus.system_services.lifecycle.zone_finalizers.cache_finalizer import CacheZoneFinalizer
+from nexus.system_services.lifecycle.zone_finalizers.mount_finalizer import MountZoneFinalizer
+from nexus.system_services.lifecycle.zone_finalizers.rebac_finalizer import ReBACZoneFinalizer
+from nexus.system_services.lifecycle.zone_finalizers.search_finalizer import SearchZoneFinalizer
 
 # ---------------------------------------------------------------------------
 # CacheZoneFinalizer
